requirements.txt
# Requisitos para el entrenamiento del modelo

keras==3.4.1
tensorflow==2.17.0
numpy==1.26.4
opencv-python==4.10.0.84
matplotlib==3.9.0
Pillow==10.3.0
requests==2.32.3


# Requisitos para uso de la api

flask==3.0.3
flask-cors==4.0.1